Title Screen + Fix Warning

+++ 05-ScenceManager/Brick.cpp
@@ -79,8 +79,8 @@ void Brick::HandleCollision(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 	// No collision occured, proceed normally
 	if (coEvents.size() == 0)
 	{
-		x += dx;
-		y += dy;
+		x += (float)dx;
+		y += (float)dy;
 	}
 	else
 	{
